+++ src/main/java/com/app/booking/internal/ticket_service/dto/request/BookRequest.java
@@ -14,5 +14,5 @@ public class BookRequest {
     String userId;
 
     @NonNull
-    Long seatId;
+    Integer seatId;
 }
