+++ src/main/java/com/app/booking/internal/ticket_service/repository/TicketRepository.java
@@ -7,6 +7,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface TicketRepository extends JpaRepository<Ticket,Long> {
+public interface TicketRepository extends JpaRepository<Ticket,Integer> {
     Page<Ticket> findAllByUserId(String userId, Pageable pageable);
 }
